Error in rule 'UpdateAsignment'.
While the description instructs &-users to populate the 'var' relation, there is (obviously) no guarantee that this is done. It is my experience that patterns/processes such as these, that are meant to be reusable, sooner or later stumble upon &-users that don't follow the instructions (they may not have read them). This would lead to unwanted behaviour of 'InsPair'.
Note that for the same reason, this rule should be split in two rules: one that updates Assignments that have a trace, and another that updates Assignments that do not have a trace. I leave this to the author.